Add continue_function handler
This function gets another function as it parameter
if the user choose the y the continue_function calls
the function passed to it to continue the actual
operation, else calls the menu. If the user choose
an invalid option the continue_function calls itself
again with the parameter passed at the first state.

Signed-off-by: Gustavo Porto Guedes <gportog@br.ibm.com>